% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeApiCall.R
\name{as.character.response}
\alias{as.character.response}
\title{S3 method to turn curl response into character}
\usage{
\method{as.character}{response}(x, ...)
}
\arguments{
\item{x}{response from curl to render to character}

\item{...}{If type='text/csv' this is passed to read.csv. If type='application/json'
this is sent to jsonlite::fromJSON}
}
\description{
Converts a raw curl response into a character string.
}
