% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRecordsTyped.R
\name{castHelper}
\alias{castHelper}
\alias{castLabel}
\alias{caseCode}
\alias{castRaw}
\alias{castChecked}
\alias{castCheckLabel}
\alias{castCheckCode}
\title{Cast REDCap data helper functions}
\usage{
castLabel(x, coding, field_name)

caseCode(x, coding, field_name)

castRaw(x, coding, field_name)

castChecked(x, coding, field_name)

castCheckLabel(x, coding, field_name)

castCheckCode(x, coding, field_name)
}
\arguments{
\item{x}{character vector. The raw data to cast.}

\item{coding}{named character vector. The defined coding from the meta data.}

\item{field_name}{character(1). Name of the field}
}
\value{
type cast vector
}
\description{
returns a type cast of a character field. Helper for
\code{\link{exportRecordsTyped}}
}
