% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changedRecords.R
\name{changedRecords}
\alias{changedRecords}
\title{get a list of records changed (adds, updates, deletes)}
\usage{
changedRecords(rcon, ...)
}
\arguments{
\item{rcon}{A redcapConnection object.}

\item{...}{Arguments passed to \code{exportLogging}}
}
\value{
Returns a list with the elements\tabular{ll}{
    \tab  \cr
   \code{updated} \tab character vector of updated record's ids \cr
   \code{deleted} \tab charcater vector of deleted record's ids \cr
   \code{created} \tab character vector of created record's ids \cr
}
}
\description{
This is a convience function that scans logs and
returns record ids of changes.
}
\details{
Makes a call to \code{exportLogging} and
}
\examples{
\dontrun{
unlockREDCap(connections = c(rcon = "project_alias"), 
             url = "your_redcap_url", 
             keyring = "API_KEYs", 
             envir = globalenv())

# Changes in last 24 hours     
changedRecords(rcon, beginTime=as.POSIXct(Sys.time()-86400))
}

}
