% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{FILE_REPOSITORY_EMPTY_FRAME}
\title{Constants for Use in the redcapAPI Package}
\format{
An object of class \code{data.frame} with 0 rows and 4 columns.
}
\usage{
FILE_REPOSITORY_EMPTY_FRAME
}
\description{
Constants that are used inside the package
}
\keyword{datasets}
