% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRedcapProject.R
\name{createRedcapProject}
\alias{createRedcapProject}
\alias{createRedcapProject.redcapApiConnection}
\title{Create REDCap Project}
\usage{
createRedcapProject(
  rcon,
  project_title,
  purpose = REDCAP_PROJECT_PURPOSE,
  purpose_other = NULL,
  is_longitudinal = FALSE,
  surveys_enabled = FALSE,
  record_autonumbering_enabled = FALSE,
  xml = NULL,
  ...
)

\method{createRedcapProject}{redcapApiConnection}(
  rcon,
  project_title,
  purpose = REDCAP_PROJECT_PURPOSE,
  purpose_other = NULL,
  is_longitudinal = FALSE,
  surveys_enabled = FALSE,
  record_autonumbering_enabled = FALSE,
  xml = NULL,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{project_title}{\code{character(1)} Title for the new project.}

\item{purpose}{\code{character}, one of 
\code{c("Practice/just for fun", "Other", "Research", "Quality Improvement", "Operational Support")}}

\item{purpose_other}{\code{character(1)} or \code{NULL}. Ignored unless 
\code{purpose = "Other"}, in which case this becomes a required argument.}

\item{is_longitudinal}{\code{logical(1)}, determines if the project will 
be set as a longitudinal project.}

\item{surveys_enabled}{\code{logical(1)}, determines if surveys are enabled
for the project. (This will not add any survey instruments, only 
enable them).}

\item{record_autonumbering_enabled}{\code{logical(1)}, determines if 
auto numbering will be enabled in the project.}

\item{xml}{\code{character(1)} an XML string in CDISC ODM XML format that 
contains project metadata (fields, forms, events, arms) and might 
optionally contain data to be imported as well. The XML contained in 
this parameter can come from a REDCap Project XML export file from 
REDCap itself, or may come from another system that is capable of 
exporting projects and data in CDISC ODM format. If the 'odm' parameter 
is included in the API request, it will use the XML to import its 
contents into the newly created project. This will allow you not only 
to create the project with the API request, but also to import all fields, 
forms, and project attributes (and events and arms, if longitudinal) as 
well as record data all at the same time.}

\item{...}{Additional arguments to be passed between methods.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to create a new REDCap project. 
  A 64-character Super API Token is required for this method (as opposed 
  to project-level API methods that require a regular 32-character 
  token associated with the project-user). In the API request, you must 
  minimally provide the project attributes 'project_title' and 
  'purpose' (with numerical value 0=Practice/Just for fun, 1=Other, 
  2=Research, 3=Quality Improvement, 4=Operational Support) when 
  creating a project.
  
  When a project is created with this method, the project will 
  automatically be given all the project-level defaults just as if you
  created a new empty project via the web user interface, such as a 
  automatically creating a single data collection instrument seeded with a 
  single Record ID field and Form Status field, as well as 
  (for longitudinal projects) one arm with one event. And if you intend 
  to create your own arms or events immediately after creating the project, 
  it is recommended that you utilize the override=1 parameter in the 
  'Import Arms' or 'Import Events' method, respectively, so that the 
  default arm and event are removed when you add your own. Also, the user 
  creating the project will automatically be added to the project as a
   user with full user privileges and a project-level API token, 
   which could then be used for subsequent project-level API requests.
   
   NOTE: Only users with Super API Tokens can utilize this method. 
   Users can only be granted a super token by a REDCap administrator 
   (using the API Tokens page in the REDCap Control Center). Please be 
   advised that users with a Super API Token can create new REDCap projects 
   via the API without any approval needed by a REDCap administrator.
   If you are interested in obtaining a super token, please contact your 
   local REDCap administrator.
}
