% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteDags.R, R/exportDags.R, R/importDags.R
\name{deleteDags}
\alias{deleteDags}
\alias{exportDags,}
\alias{deleteDags.redcapApiConnection}
\alias{exportDags}
\alias{exportDags.redcapApiConnection}
\alias{importDags}
\alias{importDags.redcapApiConnection}
\title{Export, Import, Delete Data Access Groups from a Project}
\usage{
deleteDags(rcon, dags, ...)

\method{deleteDags}{redcapApiConnection}(
  rcon,
  dags,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)

exportDags(rcon, ...)

\method{exportDags}{redcapApiConnection}(
  rcon,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)

importDags(rcon, data, ...)

\method{importDags}{redcapApiConnection}(
  rcon,
  data,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{dags}{\code{character} vector of names matching the \code{unique_group_name}.}

\item{...}{Additional arguments to pass to other methods.}

\item{refresh}{\code{logical(1)}. When \code{TRUE}, cached event data will 
be refreshed after the import.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}

\item{data}{A \code{data.frame} with two columns: \code{data_access_group_name}
and \code{unique_group_name}.}
}
\value{
\code{exportDags} with the columns
\itemize{
  \item{\code{data_access_group_name}}{The human readable name for the data access group.}
  \item{\code{unique_group_name}}{The internal unique group name}
  \item{\code{data_access_group_id}}{The internal numeric identifier.}
}

\code{importDags} has no return, but will print a message indicating the
number of Data Access Groups imported. 

\code{deleteDags} has no return, but will print a message indicating the
number of Data Access Groups deleted.
}
\description{
These methods allow you to export existing Data Access Groups, 
  import new Data Access Groups, or delete Data Access Groups from a 
  project.
}
\details{
When importing data access groups, provide a value for \code{data_access_group_name} 
  with no value for \code{unique_group_name}. 
  
  To modify a group name, provide a new value for \code{data_access_group_name}
  with the associated \code{unique_group_name}. If \code{unique_group_name}
  is provided, it must match a value currently in the project.
}
\seealso{
\code{\link{exportUserDagAssignments}}, \code{\link{importUserDagAssignments}}
}
