% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUsers.R
\name{deleteUsers}
\alias{deleteUsers}
\alias{deleteUsers.redcapApiConnection}
\title{Delete REDCap Users from a Project}
\usage{
deleteUsers(rcon, users, ...)

\method{deleteUsers}{redcapApiConnection}(
  rcon,
  users,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{users}{\code{character} vector of unique user names to be deleted.}

\item{...}{Arguments to be passed to other methods.}

\item{refresh}{\code{logical(1)} If \code{TRUE}, the cached arms data will
be refreshed after the deletion.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to delete users from a project
}
