% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRepeatingNA.R
\name{dropRepeatingNA}
\alias{dropRepeatingNA}
\title{Drop Row Where Repeat Instrment Is NA}
\usage{
dropRepeatingNA(Records, rcon, quiet = FALSE)
}
\arguments{
\item{Records}{A \code{data.frame} containing the records from
\code{\link{exportRecordsTyped}}}

\item{rcon}{A REDCap connection object as created by \code{redcapConnection}.}

\item{quiet}{logical. Suppress messages.}
}
\description{
Drops rows where the repeat instrument is NA. Returns a data frame of records where repeat instruments have a value.
}
