% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBioportalCoding.R
\name{exportBioportalCoding}
\alias{exportBioportalCoding}
\alias{exportBioportalCoding.redcapApiConnection}
\title{Export Codebook Mapping for BioPortal Fields}
\usage{
exportBioportalCoding(rcon, fields, ...)

\method{exportBioportalCoding}{redcapApiConnection}(
  rcon,
  fields = NULL,
  ...,
  batch_size = NULL,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{fields}{\code{character} or \code{NULL}. Vector of fields to be returned.  If \code{NULL},
all fields are returned (unless \code{forms} is specified).}

\item{...}{Arguments to pass to other methods}

\item{batch_size}{\code{integerish(1)} or \code{NULL}. When \code{NULL},
all records are pulled. Otherwise, the records all pulled in batches of this size.}

\item{error_handling}{\code{character(1)}. One of \code{c("error", "null")}.
An option for how to handle errors returned by the API.
see \code{\link[=redcapError]{redcapError()}}.}

\item{config}{A named \code{list}. Additional configuration parameters to pass to
\code{\link[httr:POST]{httr::POST()}}. These are appended to any parameters in
\code{rcon$config}.}

\item{api_param}{A named \code{list}. Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
Returns a named list. Each element is in the list is named for the field
it maps. The elements are named vectors where the name is the label
of the BioPortal field and the value is the code.
}
\description{
These methods enable \code{redcapAPI} to obtain a mapping of
codes and associated labels for BioPortal fields.
BioPortal fields are text fields that have enabled the
BioPortal Ontology Service Validation. The mapping between coded and
labeled values is not exported with the meta data for these fields.
Due to their size, it is impractical to attempt to maintain these tables
within \code{redcapAPI}.
}
\details{
The methods operate by executing two API calls to export first the
coded values and then the labeled values of BioPortal fields. The
two exports are then used to generate the code-label mappings for use
in casting data.
}
