% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDataQuality.R
\name{exportDataQuality}
\alias{exportDataQuality}
\title{A helper function to export data queries from the Data Quality REDCap
module.}
\usage{
exportDataQuality(rcon, prefix, ...)
}
\arguments{
\item{rcon}{A REDCap connection object as generated by \code{redcapConnection}.}

\item{prefix}{A string from your REDCap institutions Data Quality module url. The
module prefix can be found by exporting module settings under External Modules
in REDCap. At VUMC the prefix is 'vanderbilt_dataQuality'.}

\item{..., }{additional arguments that are ignored.}
}
\description{
Exports Data Quality queries by record. The Data Quality module
must be enabled on the Control Center of REDCap to use this function. Additionally,
this module must be enabled on each project before it can be used.
}
