% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSAS.R
\name{exportSAS}
\alias{exportSAS}
\title{Export the REDCap data as a SAS importable set of files.}
\usage{
exportSAS(rcon, directory = NULL, ...)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{directory}{\code{NULL} or \code{character(1)}. Directory to write files into.
Defaults to current working directory.}

\item{...}{Arguments to pass to other methods}
}
\value{
A vector of exported data set names.
}
\description{
This creates a csv for each form and a SAS script that will
load all the data into SAS.
}
\details{
This function is \emph{experimental} and needs feedback/suggestions to flesh
it out fully. If this feature is important to you, please consider
opening an issue on github to suggest improvements.

This function relies on \code{\link{exportBulkRecords}} to do the
bulk of the export before formatting for SAS. \dots are supplied to
exportBulkRecords so full user inversion of control still applies.
}
\examples{
\dontrun{
exportSAS(rcon)
}
}
\seealso{
\code{\link[=exportBulkRecords]{exportBulkRecords()}}
}
