% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSurveyLink.R
\name{exportSurveyLink}
\alias{exportSurveyLink}
\alias{exportSurveyLink.redcapApiConnection}
\title{Export Survey Link for a Participant}
\usage{
exportSurveyLink(rcon, record, instrument, event, repeat_instance = 1, ...)

\method{exportSurveyLink}{redcapApiConnection}(
  rcon,
  record,
  instrument,
  event = NULL,
  repeat_instance = 1,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{record}{\code{character(1)} giving the record ID. May also be 
numeric (will be coerced to character).}

\item{instrument}{\code{character(1)} giving the survey instrument. This 
must be one of the form names listed in the meta data.}

\item{event}{\code{character(1)} A unique event name. Only applies to 
longitudinal projects.}

\item{repeat_instance}{\code{integerish(1)}, the repeat instance if the
instrument is designated as a repeating instrument. Default value is 1.}

\item{...}{Additional arguments to pass to other methods.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method returns a unique survey link (i.e., a URL) in 
  plain text format for a specified record and data collection 
  instrument (and event, if longitudinal) in a project. If the user 
  does not have 'Survey Distribution Tools' privileges, they will not be 
  able to use this method, and an error will be returned. If the 
  specified data collection instrument has not been enabled as a 
  survey in the project, an error will be returned.
}
