% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSurveyQueueLink.R
\name{exportSurveyQueueLink}
\alias{exportSurveyQueueLink}
\alias{exportSurveyQueueLink.redcapApiConnection}
\title{Export Survey Queue Link}
\usage{
exportSurveyQueueLink(rcon, record, ...)

\method{exportSurveyQueueLink}{redcapApiConnection}(
  rcon,
  record,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{record}{\code{character(1)} giving the record ID. May also be 
numeric (will be coerced to character).}

\item{...}{Additional arguments to pass to other methods.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method returns a unique Survey Queue link (i.e., a URL) 
  in plain text format for the specified record in a project that is 
  utilizing the Survey Queue feature. If the user does not have 
  'Survey Distribution Tools' privileges, they will not be able to use 
  this method, and an error will be returned. If the Survey Queue 
  feature has not been enabled in the project, an error will be returned.
}
