% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportUserRoleAssignments.R
\name{exportUserRoleAssignments}
\alias{exportUserRoleAssignments}
\alias{exportUserRoleAssignments.redcapApiConnection}
\title{Export User-Role Assignments}
\usage{
exportUserRoleAssignments(rcon, ...)

\method{exportUserRoleAssignments}{redcapApiConnection}(
  rcon,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{...}{Arguments to be passed to other methods.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to export existing User-Role assignments for a project
}
