% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldChoiceMapping.R
\name{fieldChoiceMapping}
\alias{fieldChoiceMapping}
\alias{fieldChoiceMapping.character}
\alias{fieldChoiceMapping.redcapApiConnection}
\title{Split a Field Choice Mapping Into a Two Column Matrix}
\usage{
fieldChoiceMapping(object, field_name, ...)

\method{fieldChoiceMapping}{character}(object, field_name, ...)

\method{fieldChoiceMapping}{redcapApiConnection}(object, field_name, ...)
}
\arguments{
\item{object}{\code{redcapConnection} or \code{character(1)}. When \code{character}, is
matches the format of the meta data field choices
(i.e. \code{rcon$meta_data()$select_choices_or_calculations}).}

\item{field_name}{\code{character(1)} gives the field name for which to
make the choice mapping.}

\item{...}{Arguments to pass to other methods}
}
\value{
Returns a matrix with two columns, \code{choice_value} and \code{choice_label}
}
\description{
Uses the string from the \code{select_choices_or_calculations}
for the meta data to create a matrix of codes and their mapped
labels.
}
