% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileRepositoryPath.R
\name{fileRepositoryPath}
\alias{fileRepositoryPath}
\title{Reconstruct the file repository path}
\usage{
fileRepositoryPath(doc_id = numeric(0), folder_id = numeric(0), fileRepo)
}
\arguments{
\item{doc_id}{\code{integerish(0/1)}. The document ID for which the 
file path should be returned. Only one of \code{doc_id} or 
\code{folder_id} should be specified.}

\item{folder_id}{\code{integerish(0/1)}. The folder ID for which the 
file path should be returned. Only one of \code{doc_id} or 
\code{folder_id} should be specified.}

\item{fileRepo}{\code{data.frame} with the file repository listing. 
Typically provided by \code{rcon$fileRepository()}}
}
\description{
Reconstruct the file repository path
}
\author{
Benjamin Nutter
}
