% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterEmptyRow.R
\name{filterEmptyRow}
\alias{filterEmptyRow}
\title{Remove Rows Containing Only Missing Values}
\usage{
filterEmptyRow(data, rcon)
}
\arguments{
\item{data}{A \code{data.frame} to be filtered.}

\item{rcon}{A \code{redcapConnection} object.}
}
\description{
Evaluates each row of a data frame for missingness. If all
fields (excluding the identifying fields) are missing, the row is
removed from the data. For the purpose of this function,
\code{redcap_data_access_group} is considered an identifying field.
}
\seealso{
\code{\link[=exportRecordsTyped]{exportRecordsTyped()}}, \cr
\code{\link[=exportReportsTyped]{exportReportsTyped()}}
}
