% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessCast.R
\name{guessCast}
\alias{guessCast}
\alias{guessDate}
\title{A helper function to make a guess at casting uncast columns}
\usage{
guessCast(
  Records,
  conn,
  na = isNAorBlank,
  validation,
  cast,
  quiet = FALSE,
  threshold = 0.8
)

guessDate(
  Records,
  conn,
  na = isNAorBlank,
  validation = valRx("^[0-9]{1,4}-(0?[1-9]|1[012])-(0?[1-9]|[12][0-9]|3[01])$"),
  cast = function(x, ...) as.POSIXct(x, format = "\%Y-\%m-\%d"),
  quiet = FALSE,
  threshold = 0.8
)
}
\arguments{
\item{Records}{A \code{data.frame} containing the records from
\code{\link{exportRecordsTyped}}}

\item{na}{function. A function of the signature function(x) that determines
if a field is NA.}

\item{validation}{function. A function of the signature function(x) that
determines if a field is valid.}

\item{cast}{function. A function of the signature function(x) that
casts a column passing the validation.}

\item{threshold}{numeric(1). The threshold of non-NA data to trigger casting.}

\item{conn.}{The REDCap connection object.}
}
\description{
Will do a type cast if a validation is met above
 a threshold ratio of non-NA records. It modifies the existing
 \code{invalid} attribute to reflect the cast.
}
\examples{
\dontrun{
recs <- exportRecordsTyped(rcon, cast=raw_cast) |> 
  guessCast(validation=valRx("^[0-9]{1,4}-(0?[1-9]|1[012])-(0?[1-9]|[12][0-9]|3[01])$"), 
            cast=as.Date,
            threshold=0.6)
}
}
