% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importDags.R
\name{importDags}
\alias{importDags}
\alias{importDags.redcapApiConnection}
\title{Import Data Access Groups to a REDCap Project}
\usage{
importDags(rcon, data, ...)

\method{importDags}{redcapApiConnection}(
  rcon,
  data,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object}

\item{data}{A \code{data.frame} with two columns: \code{data_access_group_name}
and \code{unique_group_name}.}

\item{...}{Additional arguments to pass to other methods.}

\item{refresh}{\code{logical(1)}. When \code{TRUE}, cached event data will 
be refreshed after the import.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to import new DAGs (Data Access Groups) 
  into a project or update the group name of any existing DAGs.
  
  NOTE: DAGs can be renamed by simply changing the group name 
  (data_access_group_name). DAG can be created by providing group name 
  value while unique group name should be set to blank.
}
\details{
To add a DAG, provide a value for \code{data_access_group_name} 
  with value for \code{unique_group_name}. 
  
  To modify a group name, provide a new value for \code{data_access_group_name}
  with the associated \code{unique_group_name}. If \code{unique_group_name}
  is provided, it must match a value currently in the project.
}
