% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importUserRoleAssignments.R
\name{importUserRoleAssignments}
\alias{importUserRoleAssignments}
\alias{importUserRoleAssignments.redcapApiConnection}
\title{Import User-Role Assignments}
\usage{
importUserRoleAssignments(rcon, data, ...)

\method{importUserRoleAssignments}{redcapApiConnection}(
  rcon,
  data,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{data}{A \code{data.frame} with columns \code{username} and 
\code{unique_role_name}. Each \code{username} must be unique. 
Users without a \code{unique_role_name} will not be assigned to 
a user role.}

\item{...}{Arguments to be passed to other methods.}

\item{refresh}{\code{logical(1)}. When \code{TRUE}, the cached value
in \code{rcon} will be refreshed after the import.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to assign users to any user role.

NOTE: If you wish to modify an existing mapping, you must provide its 
unique username and role name. If the 'unique_role_name' column is not 
provided, user will not assigned to any user role. 
There should be only one record per username.
}
