% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid.R
\name{invalidSummary}
\alias{invalidSummary}
\alias{format.invalid}
\alias{print.invalid}
\alias{summary.invalid}
\title{Helper functions for formatting validation failure report}
\usage{
\method{format}{invalid}(x, ...)

\method{print}{invalid}(x, ...)

\method{summary}{invalid}(object, ...)
}
\arguments{
\item{x}{The \code{invalid} class object.}

\item{...}{additional arguments to \code{print}}

\item{object}{The \code{invalid} class object.}
}
\description{
\code{\link[=exportRecordsTyped]{exportRecordsTyped()}} may have an \code{invalid}
attribute if validations fail. This data has some routines which
help locate the failing records.
}
\examples{
\dontrun{
rcon <- redcapConnection(url=[YOUR_REDCAP_URL], token=[API_TOKEN])

rec <- exportRecordsTyped(rcon)

attr(rec, "invalid")
}

}
