% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRecordsTyped.R
\name{is_na_or_blank}
\alias{is_na_or_blank}
\title{Helper function for exportRecords to determine if NA or blank.}
\usage{
is_na_or_blank(x, ...)
}
\arguments{
\item{x}{character. A vector to check for NA or blank.}

\item{...}{Consumes anything else passed to function. I.e., field_name and 
coding.}
}
\value{
logical.
}
\description{
returns TRUE/FALSE if field is NA or blank. Helper
function for constructing na overrides in \code{\link{exportRecords}}.
}
