% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isZeroCodedCheckField.R
\name{isZeroCodedCheckField}
\alias{isZeroCodedCheckField}
\alias{warnOfZeroCodedCheckCasting}
\alias{warnZeroCodedFieldPresent}
\title{Identify Check Fields with a Zero Coded Option}
\usage{
isZeroCodedCheckField(field_name)

warnOfZeroCodedCheckCasting(field_name, x)

warnZeroCodedFieldPresent(field_names)
}
\arguments{
\item{field_name}{\code{character(1)} The name of a field to be tested.}

\item{x}{\code{atomic} object.}

\item{field_names}{\code{character} vector of field names.}
}
\description{
Check fields that have 0 as a coding option can confuse
  certain data processing steps because it can be difficult to 
  differentiate if a "0" value represents an unchecked or checked 
  option. Identifying these fields is important to handling them
  correctly.
}
