% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logEvent}
\alias{logEvent}
\title{Log event}
\usage{
logEvent(severity, ...)
}
\arguments{
\item{severity}{\code{string} One of the following: 'TRACE', 'DEBUG', 'INFO',
'WARN', or 'ERROR'}

\item{...}{Information to include in the log event. Each argument must
have a name.}
}
\description{
This is one of the more complex integration of services into the `redcapAPI``
package. It's purpose is to provide the ability for a system administrator
(or user) to integrate logging into a report or application. The ability
to inject a logging framework without a developer's code being altered.
}
\details{
To do this the callback function is pulled from the option \code{redcapAPI_logger}
which defaults to doing nothing.

When the package starts up, it checks to see if SPLUNK_TOKEN and SPLUNK_URL
ENV variables are set and if so, it automatically redirects the
\code{redcapAPI_logger} to point at Splunk. It will also use SPLUNK_PROJECT
if defined, otherwise the project will be the directory name that the
code is executing from.
}
\examples{
\dontrun{
  options(redcapAPI_logger=function(severity, ...) {cat(severity, ' ', dput(list(...)), '\n')})
  logEvent("INFO", "This is a logged event")
}

}
