% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mChoiceCast.R
\name{mChoiceCast}
\alias{mChoiceCast}
\title{A helper function that adds the \code{Hmisc::mChoice} multiple 
       choice class.}
\usage{
mChoiceCast(Records, rcon, style = "labelled")
}
\arguments{
\item{Records}{A \code{data.frame} containing the records from
\code{\link{exportRecordsTyped}}}

\item{rcon}{A REDCap connection object as created by \code{redcapConnection}.}

\item{style}{character. One of "labelled" or "coded". Default is "labelled"}
}
\description{
Adds a column for a multiple choice checkbox that is cast
to the \code{Hmisc::mChoice} class. Requires \code{Hmisc} to be 
loaded.
}
\examples{
\dontrun{
recs <- exportRecordsTyped(rcon) |> mChoiceCast(rcon)
}
}
