% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepUserImportData.R
\name{prepUserImportData}
\alias{prepUserImportData}
\title{Prepare User Data for Import}
\usage{
prepUserImportData(data, rcon, consolidate = TRUE)
}
\arguments{
\item{data}{\code{data.frame} that has the structure of 
\code{redcapAPI:::REDCAP_USER_STRUCTURE}. It may also have additional 
columns for the form and export access of each of the instruments.}

\item{rcon}{\code{redcapConnection}. Used to determine the instruments
in the project.}

\item{consolidate}{\code{logical(1)} If \code{TRUE}, the form and data 
export access values will be read from the expanded columns. Otherwise, 
the consolidated values (as provided by the API export) are utilized.}
}
\description{
Prepares a data frame for import via the API. Allows for 
  data to be passed in either the raw format or the labelled data 
  received from \code{exportUsers}.
}
