% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recastForImport.R
\name{recastForImport}
\alias{recastForImport}
\title{Cast Fields for Import to a REDCap Project}
\usage{
recastForImport(
  data,
  rcon,
  fields = NULL,
  na = list(),
  validation = list(),
  cast = list(),
  ...
)
}
\arguments{
\item{data}{\code{data.frame} to prepare for import.}

\item{rcon}{A \code{redcapConnection} object.}

\item{na}{A named \code{list} of user specified functions to determine if the
data is NA. This is useful when data is loaded that has coding for NA, e.g.
-5 is NA. Keys must correspond to a truncated REDCap field type, i.e.
{date_, datetime_, datetime_seconds_, time_mm_ss, time_hh_mm_ss, time, float,
number, calc, int, integer, select, radio, dropdown, yesno, truefalse,
checkbox, form_complete, sql}. The function will be provided the variables
(x, field_name, coding). The function must return a vector of logicals
matching the input. It defaults to \code{\link{isNAorBlank}} for all
entries.}

\item{validation}{A named \code{list} of user specified validation functions. The 
same named keys are supported as the na argument. The function will be 
provided the variables (x, field_name, coding). The function must return a
vector of logical matching the input length. Helper functions to construct
these are \code{\link{valRx}} and \code{\link{valChoice}}. Only fields that
are not identified as NA will be passed to validation functions.}

\item{cast}{A named \code{list} of user specified class casting functions. The
same named keys are supported as the na argument. The function will be 
provided the variables (x, field_name, coding). The function must return a
vector of logical matching the input length. See \code{\link{fieldValidationAndCasting}}}
}
\description{
This is a special case recast function for transforming 
  a data from to be imported to a REDCap project. It operates in a 
  similar fashion to \code{exportRecordsTyped} where it will scan 
  data to determine that it is in a valid format, and provide a report
  of any invalid results so that they can be investigated. The
  result is a data frame where every column is a character vector in 
  a format acceptable for API import method.
}
