% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recastRecords.R
\name{recastRecords}
\alias{recastRecords}
\title{Recast Data Fields}
\usage{
recastRecords(data, rcon, fields, cast = invert_default_cast, suffix = "")
}
\arguments{
\item{data}{\code{data.frame} with the data fields to be recoded.}

\item{rcon}{\code{recapConnection} object.}

\item{fields}{\code{character/logical/integerish}. A vector for identifying
which fields to recode. When \code{logical}, the length must match 
the number of columns in \code{data} (i.e., recycling not permitted).
A message is printed if any of the indicated fields are not a 
multiple choice field; no action will be taken on such fields.
For this function, yes/no and true/false fields are considered 
multiple choice fields. Fields of class \code{mChoice} are quietly skipped.}

\item{cast}{A named \code{list} of user specified class casting functions. 
Keys must correspond to a truncated REDCap field type, i.e.
{date_, datetime_, datetime_seconds_, time_mm_ss, time_hh_mm_ss, time, float,
number, calc, int, integer, select, radio, dropdown, yesno, truefalse,
checkbox, form_complete, sql}. The function will be 
provided the variables (x, field_name, coding). 
See \code{\link{fieldValidationAndCasting}}}

\item{suffix}{\code{character(1)}. An optional suffix to provide if 
the recoded variables should be returned as new columns. For example, 
if recoding a field \code{forklift_brand} and \code{suffix = "_labelled"}, 
the result will have one column with the coded values 
(\code{forklift_brand}) and one column with the labelled values 
(\code{forklift_brand_labelled}).}
}
\description{
Allows for recasting one or more fields using 
  the codes and labels in the Meta Data.
}
\details{
field types for which no casting function is specified will
  be returned with no changes.
  
  The default \code{cast} list in 
  \code{recastRecords} inverts the default casting of \code{exportRecordsTyped}.
}
