% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapAPI-package.R
\docType{package}
\name{redcapAPI}
\alias{redcapAPI-package}
\alias{redcapAPI}
\title{Access data, meta data, and files from REDCap using the API}
\description{
REDCap is a database development tool built on MySQL.  Visit
\url{https://projectredcap.org} for more information.  REDCap
provides an API through which data, the data dictionary, files, and
project information can be accessed.  The \code{redcapAPI} package
facilitates the use of these functions and
simplifies the work needed to prepare data for
analysis.
}
\details{
The user should also refer to their institution's REDCap API
documentation as a primary resource of what is available.  Different
versions of REDCap support different features--the user's REDCap
API documentation will address the features specific to their version
of REDCap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vubiostat/redcapAPI}
  \item Report bugs at \url{https://github.com/vubiostat/redcapAPI/issues}
}

}
\author{
\strong{Maintainer}: Shawn Garbett \email{shawn.garbett@vumc.org} (\href{https://orcid.org/0000-0003-4079-5621}{ORCID}) [contributor]

Authors:
\itemize{
  \item Benjamin Nutter \email{benjamin.nutter@gmail.com} [contributor]
  \item Jeffrey Horner
}

Other contributors:
\itemize{
  \item Savannah Obregon [contributor]
  \item Thomas Obadia [contributor]
  \item Marcus Lehr \email{marcus.j.lehr@gmail.com} [contributor]
  \item Brian High [contributor]
  \item Stephen Lane [contributor]
  \item Will Beasley [contributor]
  \item Will Gray [contributor]
  \item Nick Kennedy [contributor]
  \item Tan Hsi-Nien [contributor]
  \item Jeremy Stephens [contributor]
  \item Cole Beck [contributor]
  \item Bradley Johnson [contributor]
  \item Philip Chase [contributor]
  \item Paddy Tobias [contributor]
  \item Michael Chirico [contributor]
}

}
\keyword{internal}
