% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapDataStructure.R
\name{redcapDataStructures}
\alias{redcapDataStructures}
\alias{validateRedcapData}
\alias{REDCAP_SYSTEM_FIELDS}
\alias{REDCAP_PROJECT_PURPOSE}
\alias{REDCAP_METADATA_FIELDTYPE}
\alias{REDCAP_METADATA_VALIDATION_TYPE}
\alias{REDCAP_REPEAT_INSTRUMENT_STRUCTURE}
\title{REDCap Data Structures}
\usage{
validateRedcapData(data, redcap_data)

REDCAP_SYSTEM_FIELDS

REDCAP_PROJECT_PURPOSE

REDCAP_METADATA_FIELDTYPE

REDCAP_METADATA_VALIDATION_TYPE

REDCAP_REPEAT_INSTRUMENT_STRUCTURE
}
\arguments{
\item{data}{\code{data.frame} User provided data to be compared against
the established REDCap data structure.}

\item{redcap_data}{\code{data.frame} A data set from the \code{redcapAPI}
package to use a reference for comparing to expected data structure.}
}
\description{
Utilities for recognizing and validating data structures
for use with the REDCap API
}
