% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameRecord.R
\name{renameRecord}
\alias{renameRecord}
\alias{renameRecord.redcapApiConnection}
\title{Rename a Record in a Project}
\usage{
renameRecord(rcon, record_name, new_record_name, arm = NULL, ...)

\method{renameRecord}{redcapApiConnection}(
  rcon,
  record_name,
  new_record_name,
  arm = NULL,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{record_name}{\code{character(1)} The name of an existing record
in the project. Will also accept \code{numeric(1)} (which will be
coerced to a character).}

\item{new_record_name}{\code{character(1)} The new name to give to the
record. Will also accept \code{numeric(1)} (which will be coerced to
a character).}

\item{arm}{\code{character(1)} or \code{NULL}, an optional arm number.
If \code{NULL}, then all records with same name across all arms on
which it exists (if longitudinal with multiple arms) will be
renamed to new record name, otherwise it will rename the record
only in the specified arm.}

\item{...}{Arguments to pass to other methods.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to
\code{\link[httr]{POST}}. These are appended to any parameters in
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to rename a record from a project
in a single API request.
}
