% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitForms.R
\name{splitForms}
\alias{splitForms}
\title{Split a Data Frame into its Forms}
\usage{
splitForms(Records, rcon, envir = NULL, base = NULL, post = NULL)
}
\arguments{
\item{Records}{\code{data.frame} such as one generated by \code{exportRecords} or
\code{exportRecordsTyped}}

\item{rcon}{A \code{redcapConnection} object.}

\item{envir}{environment. The target environment for the resulting list
of \code{data.frame}s. Defaults to \code{NULL} which returns the a list.
Use \code{globalenv} to assign the global environment. Will accept a number
of the environment.}

\item{base}{\code{character(1)} giving the start of the naming scheme
for the elements of the list. By default, the names of the list will
be the form names. If this value is provided, it will follow the
format \code{base.form_name}.}

\item{post}{\code{function} to apply to each element of form data
after separating them, must be of signature \verb{function(data, rcon)}.}
}
\description{
Separates a data frame from REDCap into a list of data
frames where each form constitutes an element in the list.
}
\examples{
\dontrun{
unlockREDCap(connections = c(rcon = "project_alias"), 
             url = "your_redcap_url", 
             keyring = "API_KEYs", 
             envir = globalenv())
             
Records <- exportRecordsTyped(rcon)

splitForms(Records, rcon)
}
   
  
}
\seealso{
\subsection{Other post-processing functions}{

\code{\link[=recastRecords]{recastRecords()}}, \cr
\code{\link[=guessCast]{guessCast()}}, \cr
\code{\link[=guessDate]{guessDate()}}, \cr
\code{\link[=castForImport]{castForImport()}}, \cr
\code{\link[=mChoiceCast]{mChoiceCast()}}, \cr
\code{\link[=widerRepeated]{widerRepeated()}}
}
}
