% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringCleanup.R
\name{stringCleanup}
\alias{stringCleanup}
\alias{stripHTMLTags}
\alias{stripUnicode}
\title{Remove Undesired Characters From Strings}
\usage{
stripHTMLTags(
  x,
  tags = c("p", "br", "div", "span", "b", "font", "sup", "sub"),
  ignore.case = TRUE
)

stripUnicode(x)
}
\arguments{
\item{x}{\code{character}, vector of content to be cleaned.}

\item{tags}{\code{character}, vector of HTML tags to remove from \code{x}}

\item{ignore.case}{\code{logical(1)}, should cases be ignored when matching
patterns? Defaults to \code{TRUE}.}
}
\description{
These functions are utilities to clear undesired characters
from REDCap output.
}
