% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docsSurveyMethods.R,
%   R/exportSurveyParticipants.R
\name{surveyMethods}
\alias{surveyMethods}
\alias{exportSurveyParticipants}
\alias{exportSurveyParticipants.redcapApiConnection}
\title{Export Survey Participant Information}
\usage{
exportSurveyParticipants(rcon, instrument, event, ...)

\method{exportSurveyParticipants}{redcapApiConnection}(
  rcon,
  instrument = NULL,
  event = NULL,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{instrument}{\code{character(1)}. The name of a survey instrument.}

\item{event}{\code{character(1)} The event name of the event for which
participant information should be exported.}

\item{...}{Arguments to pass to other methods}

\item{error_handling}{\code{character(1)}. One of \code{c("error", "null")}.
An option for how to handle errors returned by the API.
see \code{\link[=redcapError]{redcapError()}}.}

\item{config}{A named \code{list}. Additional configuration parameters to pass to
\code{\link[httr:POST]{httr::POST()}}. These are appended to any parameters in
\code{rcon$config}.}

\item{api_param}{A named \code{list}. Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\value{
\code{exportSurveyParticipants} returns a data frame with the columns:\tabular{ll}{
    \tab  \cr
   \code{email} \tab The e-mail address of the participant. \cr
   \code{email_occurrence} \tab The number of times the invitation has been sent (after the next invite). \cr
   \code{identifier} \tab Participant identifier (if it exists) to match the survey response to a participant. \cr
   \code{record} \tab Record ID of the participant. \cr
   \code{invitation_sent_status} \tab Boolean value indicating if a survey invitation has been sent. \cr
   \code{invitation_send_time} \tab Date/time the survey invitation was sent. \cr
   \code{response_status} \tab Boolean value indicating if the participant has responded. \cr
   \code{survey_access_code} \tab The participant's survey access code. \cr
   \code{survey_link} \tab The participant's survey link. \cr
   \code{survey_queue_link} \tab The participants' survey queue link. \cr
}
}
\description{
These methods enable the user to export information relating
to survey participants.
}
\section{Functions}{
\itemize{
\item \code{exportSurveyParticipants()}: Export survey participants for a survey instrument.

}}
\examples{
\dontrun{
unlockREDCap(connections = c(rcon = "project_alias"), 
             url = "your_redcap_url", 
             keyring = "API_KEYs", 
             envir = globalenv())
             
# Export survey participants
exportSurveyParticipants(rcon, 
                         instrument = "survey_form")
                         
# Export survey participants for an event
exportSurveyParticipants(rcon, 
                         instrument = "survey_form", 
                         event = "event_1_arm_1")
                         
}

}
