% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRecordsTyped.R
\name{val_choice}
\alias{val_choice}
\title{Validate function for choice variables.}
\usage{
val_choice(x, field_name, coding)
}
\arguments{
\item{rx}{character. The regular expression to check.}

\item{...}{Consumes anything else passed to function. I.e., field_name and coding.}
}
\value{
logical.
}
\description{
returns TRUE/FALSE if field matches the coding. Helper for
\code{\link{exportRecords}}
}
