% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRecords.R
\name{val_rx}
\alias{val_rx}
\title{Construct a validate function from a regex.}
\usage{
val_rx(rx)
}
\arguments{
\item{rx}{character. The regular expression to check.}

\item{...}{Consumes anything else passed to function. I.e., field_name and coding.}
}
\value{
logical.
}
\description{
returns function that will validate using the given regex.
}
