% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorToApiBodyList.R
\name{vectorToApiBodyList}
\alias{vectorToApiBodyList}
\title{Convert R Vector To List for the API Call Body}
\usage{
vectorToApiBodyList(vector, parameter_name)
}
\arguments{
\item{vector}{An \code{atomic} vector.}

\item{parameter_name}{\code{character(1)}. The REDCap API parameter name.}
}
\description{
Converts an R vector to a list that will be suitable for
\code{makeApiCall}.
}
\examples{
\dontrun{
vectorToApiBodyList(1:3, "records")
}

}
