% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widerRepeated.R
\name{widerRepeated}
\alias{widerRepeated}
\title{Transform Data Into Wide Format}
\usage{
widerRepeated(Records, rcon, idvar)
}
\arguments{
\item{Records}{A \code{data.frame} containing the records from
\code{\link{exportRecordsTyped}}}

\item{rcon}{A REDCap connection object as created by \code{unlockREDCap}.}

\item{idvar}{\code{character}, Names of one or more variables in long format that identify multiple records belonging to the same group. 
These variables may also be present in wide format. (e.g., 'record_id')}
}
\description{
Converts a dataframe into wide format given the column names to expand.
}
