% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeDataForImport.R
\name{writeDataForImport}
\alias{writeDataForImport}
\title{Prepare a Data Frame for Import Through the API}
\usage{
writeDataForImport(data, csv_delimiter = ",")
}
\arguments{
\item{data}{\code{data.frame} to be imported to the API}

\item{csv_delimiter}{\code{character(1)} Delimiter used to separate fields in the generated CSV
string. Defaults to \code{","}.}
}
\description{
Converts a dataframe into a character value in the format
of a CSV for import through the API.
}
