% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deleteUserRoles.R
\name{deleteUserRoles}
\alias{deleteUserRoles}
\alias{deleteUserRoles.redcapApiConnection}
\title{Delete User Roles from a Project}
\usage{
deleteUserRoles(rcon, user_roles, ...)

\method{deleteUserRoles}{redcapApiConnection}(
  rcon,
  user_roles,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{user_roles}{\code{character} unique role names to be deleted from
the project.}

\item{...}{Arguments to be passed to other methods.}

\item{refresh}{\code{logical(1)}. If \code{TRUE}, the cached data will
be refreshed after the import.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcapError}}}

\item{config}{\code{list} Additional configuration parameters to pass to
\code{\link[httr]{POST}}. These are appended to any parameters in
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to delete User Roles from a project.
}
