% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropRepeatingNA.R
\name{dropRepeatingNA}
\alias{dropRepeatingNA}
\title{Drop Row Where Repeat Instrument Is NA}
\usage{
dropRepeatingNA(Records, rcon, quiet = FALSE)
}
\arguments{
\item{Records}{A \code{data.frame} containing the records from \code{\link[=exportRecordsTyped]{exportRecordsTyped()}}}

\item{rcon}{A \code{redcapConnection} object.}

\item{quiet}{\code{logical(1)}. When \code{FALSE}, a message is printed
indicating how many rows were in \code{Records} at the start and
completion of the subset.}
}
\description{
Drops rows where the repeat instrument is NA. Returns a data frame of records where repeat instruments have a value.
}
\seealso{
\code{\link[=exportRecordsTyped]{exportRecordsTyped()}}, \cr
\code{\link[=exportReportsTyped]{exportReportsTyped()}}
}
