% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRecordsTyped.R
\name{fieldValidationAndCasting}
\alias{fieldValidationAndCasting}
\alias{isNAorBlank}
\alias{valRx}
\alias{valChoice}
\title{Helper functions for \code{exportRecordsTyped} Validation and Casting}
\usage{
isNAorBlank(x, ...)

valRx(rx)

valChoice(x, field_name, coding)
}
\arguments{
\item{x}{\code{character}. A vector to check for NA or blank.}

\item{...}{Consumes anything else passed to function. I.e., field_name and 
coding.}

\item{rx}{\code{character}. The regular expression to check.}

\item{field_name}{\code{character}. The name of the field being checked.}

\item{coding}{\code{character}. The choices supported by the MetaData.}
}
\description{
Helper functions for \code{exportRecordsTyped} Validation and Casting
}
\details{
\code{isNAorBlank} returns TRUE/FALSE if field is NA or blank. Helper
function for constructing na overrides in \code{\link{exportRecordsTyped}}.

\code{valRx} returns a function that will validate a field using the given
regular expression. Returns a logical (TRUE/FALSE)

\code{valChoice} Validation function for choice variables. Returns 
TRUE/FALSE if fields matches the coding.
}
