% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldValidationAndCasting.R
\docType{data}
\name{fieldValidationAndCasting}
\alias{fieldValidationAndCasting}
\alias{isNAorBlank}
\alias{valRx}
\alias{valChoice}
\alias{castLabel}
\alias{castCode}
\alias{castRaw}
\alias{castChecked}
\alias{castCheckLabel}
\alias{castCheckCode}
\alias{raw_cast}
\title{Helper functions for \code{exportRecordsTyped} Validation and Casting}
\format{
An object of class \code{list} of length 19.
}
\usage{
isNAorBlank(x, ...)

valRx(rx)

valChoice(x, field_name, coding)

castLabel(x, coding, field_name)

castCode(x, coding, field_name)

castRaw(x, coding, field_name)

castChecked(x, coding, field_name)

castCheckLabel(x, coding, field_name)

castCheckCode(x, coding, field_name)

raw_cast
}
\arguments{
\item{x}{\code{character}. A vector to check.}

\item{...}{Consumes anything else passed to function. I.e., field_name and 
coding.}

\item{rx}{\code{character}. The regular expression pattern to check.}

\item{field_name}{\code{character(1)}. Name of the field(s)}

\item{coding}{named \code{character} vector. The defined coding from the meta data.}
}
\description{
This set of functions assists in validating that the content of 
  fields coming from REDCap match the MetaData, allowing for a 
  validation report to provided. The cast helpers allow for transforming
  the REDCap data into R data types and allowing the user to customize 
  the end product.
}
\details{
Functions passed to the \code{na}, \code{validation}, and
\code{cast} parameter of \code{\link{exportRecordsTyped}} all take the form
of \code{function(x, coding, field_name)}. \code{na} and \code{validation}
functions are expected to return a logical vector of the same length as the
column processed. Helper routines
are provided here for common cases to construct these functions. 

\code{isNAorBlank} returns TRUE/FALSE if field is NA or blank. Helper
function for constructing \code{na} overrides in \code{\link{exportRecordsTyped}}.

\code{valRx} constructs a validation function from a regular expression pattern. 
The function returns a TRUE/FALSE if the value matches the pattern.

\code{valchoice} constructs a validation function from a set of choices 
defined in the MetaData. The functions returns a TRUE/FALSE if the value
matches one of the choices.

\code{castLabel} constructs a casting function for multiple choice variables. 
The field will be cast to return the choice label (generally more human readable)

\code{castCode} constructs a casting function for multiple choice variables.
Similar to \code{castLabel}, but the choice value is returned instead. The
values are typically more compact and their meaning may not be obvious.

\code{castRaw} constructs a casting function that returns the content
from REDCap as it was received. It is functionally equivalent to \code{identity}

\code{castChecked} constructs a casting function for checkbox fields. It
returns values in the form of Unchecked/Checked.

\code{castCheckLabel} and \code{castCheckCode} also construct casting functions
for checkbox fields. For both, unchecked variables are cast to an empty 
string (""). Checked variables are cast to the option label and option code, 
respectively.

\code{raw_cast} overrides all casting if passed as the \code{cast}
parameter.
}
\author{
Shawn Garbett, Benjamin Nutter
}
\keyword{datasets}
