% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectIdFields.R
\name{getProjectIdFields}
\alias{getProjectIdFields}
\title{Return a vector of the Project ID Fields}
\usage{
getProjectIdFields(rcon)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}
}
\description{
Returns a character vector listing the project ID fields. 
  This will be at most a vector of length two. The first element will
  be the first field in the meta data. The second, if provided, will
  be the name of the secondary unique field specified in the project.
}
