% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importUsers.R
\name{importUsers}
\alias{importUsers}
\alias{importUsers.redcapApiConnection}
\title{Import Users Data to REDCap}
\usage{
importUsers(rcon, data, ...)

\method{importUsers}{redcapApiConnection}(
  rcon,
  data,
  consolidate = TRUE,
  refresh = TRUE,
  ...,
  error_handling = getOption("redcap_error_handling"),
  config = list(),
  api_param = list()
)
}
\arguments{
\item{rcon}{A \code{redcapConnection} object.}

\item{data}{A \code{data.frame} with the user data for import.}

\item{...}{Arguments to be passed to other methods.}

\item{consolidate}{\code{logical(1)} If \code{TRUE}, the form and data 
export access values will be read from the expanded columns. Otherwise, 
the consolidated values (as provided by the API export) are utilized.}

\item{refresh}{\code{logical(1)}. If \code{TRUE}, the cached data will
be refreshed after the import.}

\item{error_handling}{An option for how to handle errors returned by the API.
see \code{\link{redcap_error}}}

\item{config}{\code{list} Additional configuration parameters to pass to 
\code{\link[httr]{POST}}. These are appended to any parameters in 
\code{rcon$config}.}

\item{api_param}{\code{list} Additional API parameters to pass into the
body of the API call. This provides users to execute calls with options
that may not otherwise be supported by \code{redcapAPI}.}
}
\description{
This method allows you to import new users into a project 
while setting their user privileges, or update the privileges of 
existing users in the project.
}
\details{
When setting permissions for a user project access fields (those
  not related to forms or exports) are set as either 0 or 1 (or "No Access"
  and "Access", respectively). The settings may be any of these four values.
  
For form access fields, the values must be one of 0, 2, 1, or 3 ("No Access", 
"Read Only", "View records/responses and edit records (survey responses are read-only)", 
or "Edit survey responses", respectively). 

Data export fields must be one of 0, 2, 3, or 1 ("No Access", "De-Identified", 
"Remove Identifier Fields", "Full Data Set", respectively). 

The user data is passed through \code{prepUserImportData} before sending it
to the API, so text values listed above may also be used and will be 
converted to the numeric equivalent. 

It is also permissible to use a column for each form individually, as can
be exported via \code{exportUsers}. With \code{consolidate = TRUE}, these 
settings will be consolidated into the text string expected by the API.
}
