% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapDataStructure.R
\name{redcapDataStructures}
\alias{redcapDataStructures}
\alias{validateRedcapData}
\title{REDCap Data Structures}
\usage{
validateRedcapData(data, redcap_data)
}
\description{
Utilities for recognizing and validating data structures 
  for use with the REDCap API
}
